Functions

Table of Contents

CrossSectionModule functions

CompositeLayupModule functions

FESetShellT3Module functions

MeshFrameMemberModule functions

FEMMCorotBeamModule functions

FEMMShellT3FFModule functions

FEMMShellT3FFCompModule functions

FEMMPointMassModule functions

FEMMPointGroundedSpringModule functions

RotUtilModule functions

VisUtilModule functions

TransformerModule functions

CrossSectionModule functions

CompositeLayupModule functions

FinEtoolsFlexStructures.CompositeLayupModule.cartesian_csys — Method

cartesian_csys(axes)

Create a material Cartesian coordinate system.

axes = tuple of signed labels of the axes. For instance, (1, 2, 3) creates a coordinate system identical to the global cartesian coordinate system. (2, 1, -3) creates a coordinate system so that the first material basis vector is along the second global basis vector, the second material basis vector is along the first global basis vector, and the third material basis vector is opposite to the third global basis vector.

FinEtoolsFlexStructures.CompositeLayupModule.lamina_material — Method

lamina_material(E1, E2, nu12, G12, G13, G23)

Create a transversely isotropic lamina material with default (zero) mass density.

FinEtoolsFlexStructures.CompositeLayupModule.lamina_material — Method

lamina_material(rho, E1, E2, nu12, G12, G13, G23)

Create a transversely isotropic lamina material.

FinEtoolsFlexStructures.CompositeLayupModule.lamina_material — Method

lamina_material(rho, E, nu)

Create an isotropic lamina material.

FinEtoolsFlexStructures.CompositeLayupModule.lamina_material — Method

lamina_material(E, nu)

Create an isotropic lamina material with default (zero) mass density.

FinEtoolsFlexStructures.CompositeLayupModule.laminate_stiffnesses! — Method

laminate_stiffnesses!(cl::CompositeLayup, A, B, C)

Compute the laminate stiffnesses, membrane, extension-bending coupling, and bending.

FinEtoolsFlexStructures.CompositeLayupModule.laminate_transverse_stiffness! — Method

laminate_transverse_stiffness!(cl::CompositeLayup, H)

Computed the laminate transverse stiffness.

FinEtoolsFlexStructures.CompositeLayupModule.plane_stress_T_matrix! — Method

plane_stress_T_matrix!(Tm::Array{T, 2}, angle) where {T}

Compute the transformation matrix FROM stress vector components in the PLY coordinate system TO the LAYOUT coordinate system.

angle = angle (in radians) between the first basis vector of the layup coordinate system and the first basis vector of the ply coordinate system m, n = cosine and sine of the angle

The nomenclature is from Barbero, Finite element analysis of composite materials using Abaqus (2013).

FinEtoolsFlexStructures.CompositeLayupModule.plane_stress_Tbar_matrix! — Method

plane_stress_Tbar_matrix!(Tm::Array{T, 2}, angle) where {T}
plane_stress_Tbar_matrix!(Tm::Array{T, 2}, m, n) where {T}

Compute the transformation matrix FROM engineering strain components on the LAYUP coordinate system TO the PLY coordinate system.

angle = angle (in radians) between the first basis vector of the layup coordinate system and the first basis vector of the ply coordinate system m, n = cosine and sine of the angle

The nomenclature is from Barbero, Finite element analysis of composite materials using Abaqus (2013).

FinEtoolsFlexStructures.CompositeLayupModule.plane_stress_Tinv_matrix! — Method

plane_stress_Tinv_matrix!(Tinvm::Array{T, 2}, angle) where {T}
plane_stress_Tinv_matrix!(Tinvm::Array{T, 2}, m, n) where {T}

Compute the transformation matrix FROM stress vector components on the LAYOUT coordinate system TO the PLY coordinate system.

angle = angle (in radians) between the first basis vector of the layup coordinate system and the first basis vector of the ply coordinate system m, n = cosine and sine of the angle

The nomenclature is from Barbero, Finite element analysis of composite materials using Abaqus (2013).

FinEtoolsFlexStructures.CompositeLayupModule.thickness — Method

thickness(cl::CompositeLayup)

Compute the thickness of the layup (some of the thicknesses of the plies).

FinEtoolsFlexStructures.CompositeLayupModule.transverse_shear_T_matrix! — Method

transverse_shear_T_matrix!(Tm::Array{T, 2}, angle) where {T}
transverse_shear_T_matrix!(Tm::Array{T, 2}, m, n) where {T}

Compute the transformation matrix for the transverse shear stresses FROM the LAYOUT coordinate system TO the PLY coordinate system.

angle = angle (in radians) between the first basis vector of the layup coordinate system and the first basis vector of the ply coordinate system m, n = cosine and sine of the angle

FESetShellT3Module functions

MeshFrameMemberModule functions

FinEtoolsFlexStructures.MeshFrameMemberModule.frame_member — Method

frame_member(xyz, nL, crosssection; label = 0)

Mesh of a generally curved beam member given by the location of the vertices of the spline curve.

nL = number of elements

FinEtoolsFlexStructures.MeshFrameMemberModule.fuse_members — Method

fuse_members(members; tolerance = 0.001)

Fuse members by merging the meshes for the members.

Computes an array of meshes.

FinEtoolsFlexStructures.MeshFrameMemberModule.merge_members — Method

merge_members(members; tolerance = 0.001)

Merge together meshes of the members.

Uses fuse_members to merge the nodes, then concatenates the finite elements.

FEMMCorotBeamModule functions

FinEtoolsFlexStructures.FEMMCorotBeamModule.current_local_frame! — Method

current_local_frame!(Ft, xt, FtI, FtJ)

Compute the current local frame from the positions of the nodes and the orientation matrices (frames) at the nodes.

FinEtoolsFlexStructures.FEMMCorotBeamModule.distribloads_global — Method

distribloads_global(self::FEMMCorotBeam, geom0::NodalField{FFlt}, u1::NodalField{T}, Rfield1::NodalField{T}, dchi::NodalField{T}, fi) where {T<:Number}

Compute the load vector due to distributed loads.

Compute the global load vector corresponding to applied distributed load. Here it means force per unit length of the beam, in the configuration u1, Rfield1. These are only forces, not moments.

Note

The force intensity must be uniform across the entire element. The force intensity is given in the global coordinates.

FinEtoolsFlexStructures.FEMMCorotBeamModule.geostiffness — Method

geostiffness(self::FEMMCorotBeam, assembler::ASS, geom0::NodalField{FFlt}, u1::NodalField{T}, Rfield1::NodalField{T}, dchi::NodalField{T}) where {ASS<:AbstractSysmatAssembler, T<:Number}

Compute the geometric stiffness matrix.

FinEtoolsFlexStructures.FEMMCorotBeamModule.gyroscopic — Method

gyroscopic(self::FEMMCorotBeam,  assembler::A,
  geom::NodalField{FFlt},
  u::NodalField{T}) where {A<:AbstractSysmatAssembler, T<:Number}

Compute the quadratic-inertial-term (gyroscopic) mass matrix

This is a general routine for the abstract linear-deformation FEMM.

FinEtoolsFlexStructures.FEMMCorotBeamModule.local_cartesian_to_natural! — Method

local_cartesian_to_natural!(aN, L)

Compute transformation from local Cartesian displacements to natural deformations of a beam element.

Matrix defined in Eq (4.8) of COMPUTER METHODS IN APPLIED MECHANICS AND ENGINEERING 14 (1978) 401-451 ON LARGE DISPLACEMENT-SMALL STRAIN ANALYSIS OF STRUCTURES WITH ROTATIONAL DEGREES OF FREEDOM. J.H. ARGYRIS, P.C. DUNNE and D.W. SCHARPF

Arguments

L= current length of the element

Outputs

aN= transformation matrix to take Cartesian (local) displacement increments in the element frame and to produce increments of natural deformations; see local_frame_and_def! for the definition of the natural deformations

FinEtoolsFlexStructures.FEMMCorotBeamModule.local_forces! — Method

local_forces!(FL, PN, L, aN)

Compute forces through which the element acts on the nodes in the local coordinate system.

Arguments

PN = column vector of natural forces; L= current length of the element, aN= transformation matrix to take Cartesian (local) displacement increments in the element frame and to produce increments of natural deformations; see local_frame_and_def! for the definition of the natural deformations

Outputs

FL = vector of forces acting on the nodes in the local coordinate system

FinEtoolsFlexStructures.FEMMCorotBeamModule.local_frame_and_def! — Method

local_frame_and_def!(Ft, dN, F0, FtI, FtJ, x0, x1x2_vector, xt, RI, RJ)

Compute the current length of the element, the current element frame, and the natural deformations

Arguments

x0= array of node coordinates, one node per row, in initial configuration x1x2_vector= vector that lies in the x1-x2 local element coordinate plane, in initial configuration xt= array of node coordinates, one node per row, in initial configuration RI,RJ=nodal rotation (orthogonal) matrix

Outputs

Lt= current length of the element, Ft= current element frame (orthogonal rotation matrix) whose columns are unit vectors: they are centered halfway between the current locations of the nodes, vector 1 points from node I to node J, vector 3 is orthogonal to the sum of the nodal cross-section frame vectors 2 dN= vector of natural deformations; dN(1)= total change in length between configurations 0 and t; dN(2)= symmetric bending; dN(3)= anti-symmetric bending; dN(4)= symmetric bending dN(5)= anti-symmetric bending; dN(6)=total axial torsion angle.

FinEtoolsFlexStructures.FEMMCorotBeamModule.local_geometric_stiffness! — Method

local_geometric_stiffness!(SM, A, I2, I3, PN, L)

Compute the local geometric stiffness matrix.

Arguments

A= cross-sectional area, I2, I3=central moment of inertia of the cross-section about the x2 and x3 coordinate axis, PN= vector of natural forces; see natural_forces() for definitions L= current length of the element,

Outputs

SM = local geometric stiffness matrix, 12 x 12

This geometric stiffness matrix this consistent with relationship between the natural deformations and the natural forces that assumes there is only a linear constitutive link: no non-constitutive effects (bowing etc.) are included. This form of the geometric matrix was derived by Krenk. @BOOK{Krenk:2009, AUTHOR = {S. Krenk}, TITLE = {Non-linear Modeling and Analysis of Solids and Structures }, PUBLISHER = {Cambridge University Press}, YEAR = {2009}, isbn = {9780521830546} }

FinEtoolsFlexStructures.FEMMCorotBeamModule.local_mass! — Method

local_mass!(MM, A, I1, I2, I3, rho, L, mass_type)

Mass matrix of the beam.

Arguments

A= cross-sectional area, I1=central moment of inertia of the cross-section about the x1 axis, I2, I3=central moment of inertia of the cross-section about the x2 and x3 coordinate axis, rho=mass density, L= initial length of the element,

Outputs

MM = local mass matrix, 12 x 12 In the element frame the mass matrix is constant.

FinEtoolsFlexStructures.FEMMCorotBeamModule.local_stiffness! — Method

local_stiffness!(SM, E, G, A, I2, I3, J, A2s, A3s, L, aN, DN)

Compute the local elastic stiffness matrix.

Arguments

E, G= Young's and shear modulus, A= cross-sectional area, I2, I3=central moment of inertia of the cross-section about the x2 and x3 coordinate axis, J=St Venant torsion constant, L= current length of the element,

Outputs

SM = local stiffness matrix, 12 x 12

FinEtoolsFlexStructures.FEMMCorotBeamModule.mass — Method

mass(self::FEMMCorotBeam,  assembler::A,
  geom::NodalField{FFlt},
  u::NodalField{T}) where {A<:AbstractSysmatAssembler, T<:Number}

Compute the consistent mass matrix

This is a general routine for the abstract linear-deformation FEMM.

FinEtoolsFlexStructures.FEMMCorotBeamModule.natural_forces! — Method

natural_forces!(PN, E, G, A, I2, I3, J, A2s, A3s, L, dN, DN)

Compute the natural forces from the natural deformations.

Argument

E, G= Young's and shear modulus, A= cross-sectional area, I2, I3=central moment of inertia of the cross-section about the x2 and x3 coordinate axis, J=St Venant torsion constant, L= current length of the element, dN= column vector of natural deformations; see local_frames()

Outputs

PN = column vector of natural forces; PN[1]= axial force; PN[2]= symmetric bending moment in the plane x1-x2; PN[3]= anti-symmetric bending bending moment in the plane x1-x2; PN[4]= symmetric bending bending moment in the plane x1-x3; PN[5]= anti-symmetric bending bending moment in the plane x1-x3; PN[6]= axial torque.

FinEtoolsFlexStructures.FEMMCorotBeamModule.natural_stiffness_Bernoulli! — Method

natural_stiffness_Bernoulli!(DN, E, G, A, I2, I3, J, A2s, A3s, L)

Compute the natural stiffness matrix.

function DN = natural_stiffness(self, E, G, A, I2, I3, J, L)

Arguments

E, G= Young's and shear modulus,

A= cross-sectional area,

I2, I3= central moment of inertia of the cross-section about the x2 and x3 coordinate axis,

J= St Venant torsion constant,

A2s = effective area for shear in the direction of x2 (ignored)

A3s = effective area for shear in the direction of x3 (ignored)

L= current length of the element,

Outputs

DN = 6 x 6 natural stiffness matrix

FinEtoolsFlexStructures.FEMMCorotBeamModule.natural_stiffness_Timoshenko! — Method

natural_stiffness_Timoshenko!(DN, E, G, A, I2, I3, J, A2s, A3s, L)

Compute the natural stiffness matrix.

function DN = natural_stiffness(self, E, G, A, I2, I3, J, L)

Arguments

E, G= Young's and shear modulus,

A= cross-sectional area,

I2, I3= central moment of inertia of the cross-section about the x2 and x3 coordinate axis,

J= St Venant torsion constant,

A2s = effective area for shear in the direction of x2

A3s = effective area for shear in the direction of x3

L= current length of the element,

Outputs

DN = 6 x 6 natural stiffness matrix

FinEtoolsFlexStructures.FEMMCorotBeamModule.restoringforce — Method

restoringforce(self::FEMMCorotBeam, assembler::ASS, geom0::NodalField{FFlt}, u1::NodalField{T}, Rfield1::NodalField{T}, dchi::NodalField{T}) where {ASS<:AbstractSysmatAssembler, T<:Number}

Compute the vector of the restoring elastic forces

FinEtoolsFlexStructures.FEMMCorotBeamModule.stiffness — Method

stiffness(self::FEMMCorotBeam, assembler::ASS, geom0::NodalField{FFlt}, u1::NodalField{T}, Rfield1::NodalField{T}, dchi::NodalField{T}) where {ASS<:AbstractSysmatAssembler, T<:Number}

Compute the material stiffness matrix.

FEMMShellT3FFModule functions

FinEtools.FEMMBaseModule.associategeometry! — Method

associategeometry!(self::FEMMShellT3FF,  geom::NodalField{FFlt})

Associate geometry with the FEMM.

In this case it means evaluate the nodal normals.

FinEtools.FEMMBaseModule.inspectintegpoints — Method

inspectintegpoints(self::AbstractFEMMDeforLinear,
  geom::NodalField{FFlt},  u::NodalField{TI},
  dT::NodalField{FFlt},
  felist::FIntVec,
  inspector::F,  idat, quantity=:Cauchy;
  context...) where {TI<:Number, F<:Function}

Inspect integration point quantities.

geom - reference geometry field

u - displacement+rotation field

dT - temperature difference field

felist - indexes of the finite elements that are to be inspected: The fes to be included are: fes[felist].

context - structure: see the update!() method of the material.

inspector - function with the signature idat = inspector(idat, j, conn, x, out, loc); where idat - a structure or an array that the inspector may use to maintain some state, for instance minimum or maximum of stress, j is the element number, conn is the element connectivity, out is the output of the update!() method, loc is the location of the integration point in the reference configuration.

Return

The updated inspector data is returned.

FinEtoolsFlexStructures.FEMMShellT3FFModule.make — Method

make(integdomain, mcsys, material)

Make a T3FF FEMM from the integration domain, a coordinate system to define the orientation of the normals, and a material.

FinEtoolsFlexStructures.FEMMShellT3FFModule.make — Method

make(integdomain, mcsys, material)

Make a T3FF FEMM from the integration domain, and a material. Default isoparametric method for computing the normals is used.

FinEtoolsFlexStructures.FEMMShellT3FFModule.mass — Method

mass(self::FEMMShellT3FF,  assembler::A,  geom::NodalField{FFlt}, dchi::NodalField{TI}) where {A<:AbstractSysmatAssembler, TI<:Number}

Compute the diagonal (lumped) mass matrix

The mass matrix can be expected to be non-singular.

FinEtoolsFlexStructures.FEMMShellT3FFModule.num_normals — Method

num_normals(self::FEMMShellT3FF)

Compute the summary of the nodal normals.

FinEtoolsFlexStructures.FEMMShellT3FFModule.stiffness — Method

stiffness(self::FEMMShellT3FF, assembler::ASS, geom0::NodalField{FFlt}, u1::NodalField{TI}, Rfield1::NodalField{TI}, dchi::NodalField{TI}) where {ASS<:AbstractSysmatAssembler, TI<:Number}

Compute the material stiffness matrix.

FEMMShellT3FFCompModule functions

FinEtools.FEMMBaseModule.associategeometry! — Method

associategeometry!(self::FEMMShellT3FFComp,  geom::NodalField{FFlt})

Associate geometry with the FEMM.

In this case it means evaluate the nodal normals.

FinEtools.FEMMBaseModule.inspectintegpoints — Method

inspectintegpoints(self::AbstractFEMMDeforLinear,
  geom::NodalField{FFlt},  u::NodalField{TI},
  dT::NodalField{FFlt},
  felist::FIntVec,
  inspector::F,  idat, quantity=:Cauchy;
  context...) where {TI<:Number, F<:Function}

Inspect integration point quantities.

geom - reference geometry field

u - displacement+rotation field

dT - temperature difference field

felist - indexes of the finite elements that are to be inspected: The fes to be included are: fes[felist].

context - structure: see the update!() method of the material.

inspector - function with the signature idat = inspector(idat, j, conn, x, out, loc); where idat - a structure or an array that the inspector may use to maintain some state, for instance minimum or maximum of stress, j is the element number, conn is the element connectivity, out is the output of the update!() method, loc is the location of the integration point in the reference configuration.

Return

The updated inspector data is returned.

FinEtoolsFlexStructures.FEMMShellT3FFCompModule.make — Method

make(integdomain, layup)

Make a T3FFComp FEMM from the integration domain and a composite layup.

FinEtoolsFlexStructures.FEMMShellT3FFCompModule.mass — Method

mass(self::FEMMShellT3FFComp,  assembler::A,  geom::NodalField{FFlt}, dchi::NodalField{TI}) where {A<:AbstractSysmatAssembler, TI<:Number}

Compute the diagonal (lumped) mass matrix

The mass matrix can be expected to be non-singular.

FinEtoolsFlexStructures.FEMMShellT3FFCompModule.num_normals — Method

num_normals(self::FEMMShellT3FFComp)

Compute the summary of the nodal normals.

FinEtoolsFlexStructures.FEMMShellT3FFCompModule.stiffness — Method

stiffness(self::FEMMShellT3FFComp, assembler::ASS, geom0::NodalField{FFlt}, u1::NodalField{TI}, Rfield1::NodalField{TI}, dchi::NodalField{TI}) where {ASS<:AbstractSysmatAssembler, TI<:Number}

Compute the material stiffness matrix.

FEMMPointMassModule functions

FinEtoolsFlexStructures.FEMMPointMassModule.mass — Method

mass(self::FEMMCorotBeam,  assembler::A,
  geom::NodalField{FFlt},
  u::NodalField{T}) where {A<:AbstractSysmatAssembler, T<:Number}

Compute the consistent mass matrix

This is a general routine for the abstract linear-deformation FEMM.

FEMMPointGroundedSpringModule functions

FinEtoolsFlexStructures.FEMMPointGroundedSpringModule.stiffness — Method

mass(self::FEMMCorotBeam,  assembler::A,
  geom::NodalField{FFlt},
  u::NodalField{T}) where {A<:AbstractSysmatAssembler, T<:Number}

Compute the stiffness matrix

RotUtilModule functions

FinEtoolsFlexStructures.RotUtilModule.initial_Rfield — Method

initial_Rfield(fens)

Set up initial rotation field (rotation matrices are all identities)

FinEtoolsFlexStructures.RotUtilModule.update_rotation_field! — Method

update_rotation_field!(Rfield, dchi)

Update rotation field by exponential incremental rotation.

VisUtilModule functions

Modules = [VisualStructures]
Private = true
Order = [:function]

TransformerModule functions