Types

CrossSectionModule

FinEtoolsFlexStructures.CrossSectionModule.CrossSectionCircle — Method

CrossSectionCircle(radius, x1x2_vector, shear_resisting_area_factor = 6/7; label = 0)

Create a circular section for use with a Timoshenko model.

Commonly accepted value is shear_resisting_area_factor = 6/7.

FinEtoolsFlexStructures.CrossSectionModule.CrossSectionCircle — Method

CrossSectionCircle(radius, x1x2_vector; label = 0)

Create a circular section for use with a Bernoulli model.

FinEtoolsFlexStructures.CrossSectionModule.CrossSectionRectangle — Method

CrossSectionRectangle(d2f, d3f, x1x2_vector; label = 0)

Create a rectangular section for use with a Timoshenko model.

Commonly accepted value is shear_resisting_area_factor = 5/6.

FinEtoolsFlexStructures.CrossSectionModule.CrossSectionRectangle — Method

CrossSectionRectangle(d2f, d3f, x1x2_vector; label = 0)

Create a rectangular section for use with a Bernoulli model.

FEMMShellT3FFModule

FinEtoolsFlexStructures.FEMMShellT3FFModule.FEMMShellT3FF — Type

FEMMShellT3FF{S<:AbstractFESet, F<:Function} <: AbstractFEMM

Class for the finite element modeling machine of the T3 triangular Flat-Facet shell with the Discrete Shear Gap technology and a consistent handling of the normals.

With averaging of the transverse strain-displacement matrix or averaging of the transverse shear stiffness matrix to provide isotropic transverse shear response.

Also, the formulation is developed to correctly handle the coupling of twisting moments and transverse shear (such as in the twisted beam or the Raasch hook problems) by incorporating "nodal" normals.

Some of the programming developed consistently with the paper

[1] Cui et al, Analysis of plates and shells using an edge-based smoothed finite element method, Comput Mech (2010) 45:141–156 DOI 10.1007/s00466-009-0429-9

The stabilization factor of the shear term of

[2] Mikko Lyly, Rolf Stenberg and Teemu Vihinen, A stable bilinear element for the Reissner-Mindlin plate model Computer Methods in Applied Mechanics and Engineering 110 (1993) 343-357

is incorporated. Refer to expressions (3.12) and (3.13).

The treatment of the transformation between the element and nodal coordinates is carried out using a clean alternative to the publication

[3] Finite Elements in Analysis and Design 30 (1998) 235—242 The treatment of shell normals in ﬁnite element analysis Richard H. MacNeal, Charles T. Wilson, Robert L. Harder, Claus C. Hoﬀ The MacNeal-Schwendler Corporation, 815 Colorado Blvd., Los Angeles, CA 90041, USA

The formula for the element to nodal basis transformation is derived from the expression

[theta]_n = [A]_E^T [theta]_e + [A_3,1:3]_E^T [alpha_3]_e

by disconnecting the drilling degree of freedom from the bending degrees of freedom in the nodal basis.

The following features are incorporated to deal with nodal normals:

Nodal normals are averages of the normals of elements that meet at a node.

A crease in the surface is taken into account. In that case the normal are not averaged across the crease. At the nodes along the crease every element uses the normal to its surface instead of the nodal normal.

Configuration:

These attributes of the FEMM can be set after it's been created.

transv_shear_formulation: which formulation for the transverse shear stiffness?

FEMMShellT3FFModule.__TRANSV_SHEAR_FORMULATION_AVERAGE_B - averaged strains

FEMMShellT3FFModule.__TRANSV_SHEAR_FORMULATION_AVERAGE_K - averaged stiffness

drilling_stiffness_scale: multiplier of the generalized stiffness coefficient

threshold_angle: angle in degrees. If a nodal normal subtends angle bigger then this threshold, the nodal normal at that note is marked as invalid.

mult_el_size: multiplier of the square of the element size, used to control transverse shear stiffness.

FinEtoolsFlexStructures.FEMMShellT3FFModule.FEMMShellT3FF — Method

FEMMShellT3FF(integdomain::IntegDomain{S, F}, mcsys::CSys, material::M) where {S<:AbstractFESet, F<:Function, M}

Constructor of the T3FF shell FEMM.