Functions

CrossSectionModule

FEMMShellT3FFModule

FinEtools.FEMMBaseModule.associategeometry! — Method

associategeometry!(self::FEMMShellT3FF,  geom::NodalField{FFlt})

Associate geometry with the FEMM.

In this case it means evaluate the nodal normals.

FinEtools.FEMMBaseModule.inspectintegpoints — Method

inspectintegpoints(self::AbstractFEMMDeforLinear,
  geom::NodalField{FFlt},  u::NodalField{TI},
  dT::NodalField{FFlt},
  felist::FIntVec,
  inspector::F,  idat, quantity=:Cauchy;
  context...) where {TI<:Number, F<:Function}

Inspect integration point quantities.

geom - reference geometry field

u - displacement+rotation field

dT - temperature difference field

felist - indexes of the finite elements that are to be inspected: The fes to be included are: fes[felist].

context - structure: see the update!() method of the material.

inspector - function with the signature idat = inspector(idat, j, conn, x, out, loc); where idat - a structure or an array that the inspector may use to maintain some state, for instance minimum or maximum of stress, j is the element number, conn is the element connectivity, out is the output of the update!() method, loc is the location of the integration point in the reference configuration.

Return

The updated inspector data is returned.

FinEtoolsFlexStructures.FEMMShellT3FFModule.make — Method

make(integdomain, mcsys, material)

Make a T3FF FEMM from the integration domain, a coordinate system to define the orientation of the normals, and a material.

FinEtoolsFlexStructures.FEMMShellT3FFModule.mass — Method

mass(self::FEMMShellT3FF,  assembler::A,  geom::NodalField{FFlt}, dchi::NodalField{TI}) where {A<:AbstractSysmatAssembler, TI<:Number}

Compute the diagonal (lumped) mass matrix

The mass matrix can be expected to be non-singular.

FinEtoolsFlexStructures.FEMMShellT3FFModule.num_normals — Method

num_normals(self::FEMMShellT3FF)

Compute the summary of the nodal normals.

FinEtoolsFlexStructures.FEMMShellT3FFModule.stiffness — Method

stiffness(self::FEMMShellT3FF, assembler::ASS, geom0::NodalField{FFlt}, u1::NodalField{TI}, Rfield1::NodalField{TI}, dchi::NodalField{TI}) where {ASS<:AbstractSysmatAssembler, TI<:Number}

Compute the material stiffness matrix.