Functions

Table of Contents

CrossSectionModule functions

FESetShellT3Module functions

MeshFrameMemberModule functions

FEMMCorotBeamModule functions

FEMMShellT3FFModule functions

FEMMPointMassModule functions

FEMMPointGroundedSpringModule functions

RotUtilModule functions

VisUtilModule functions

TransformerModule functions

CrossSectionModule functions

FESetShellT3Module functions

MeshFrameMemberModule functions

FEMMCorotBeamModule functions

FinEtoolsFlexStructures.FEMMCorotBeamModule.distribloads_global — Method

distribloads_global(self::FEMMCorotBeam, geom0::NodalField{FFlt}, u1::NodalField{T}, Rfield1::NodalField{T}, dchi::NodalField{T}, fi) where {T<:Number}

Compute the load vector due to distributed loads.

Compute the global load vector corresponding to applied distributed load. Here it means force per unit length of the beam, in the configuration u1,Rfield1.

Note: the force intensity must be uniform across the entire element. Note: the force intensity is given in the global coordinates.

FinEtoolsFlexStructures.FEMMCorotBeamModule.geostiffness — Method

geostiffness(self::FEMMCorotBeam, assembler::ASS, geom0::NodalField{FFlt}, u1::NodalField{T}, Rfield1::NodalField{T}, dchi::NodalField{T}) where {ASS<:AbstractSysmatAssembler, T<:Number}

Compute the geometric stiffness matrix.

FinEtoolsFlexStructures.FEMMCorotBeamModule.gyroscopic — Method

gyroscopic(self::FEMMCorotBeam,  assembler::A,
  geom::NodalField{FFlt},
  u::NodalField{T}) where {A<:AbstractSysmatAssembler, T<:Number}

Compute the quadratic-inertial-term (gyroscopic) mass matrix

This is a general routine for the abstract linear-deformation FEMM.

FinEtoolsFlexStructures.FEMMCorotBeamModule.mass — Method

mass(self::FEMMCorotBeam,  assembler::A,
  geom::NodalField{FFlt},
  u::NodalField{T}) where {A<:AbstractSysmatAssembler, T<:Number}

Compute the consistent mass matrix

This is a general routine for the abstract linear-deformation FEMM.

FinEtoolsFlexStructures.FEMMCorotBeamModule.restoringforce — Method

restoringforce(self::FEMMCorotBeam, assembler::ASS, geom0::NodalField{FFlt}, u1::NodalField{T}, Rfield1::NodalField{T}, dchi::NodalField{T}) where {ASS<:AbstractSysmatAssembler, T<:Number}

Compute the vector of the restoring elastic forces

FinEtoolsFlexStructures.FEMMCorotBeamModule.stiffness — Method

stiffness(self::FEMMCorotBeam, assembler::ASS, geom0::NodalField{FFlt}, u1::NodalField{T}, Rfield1::NodalField{T}, dchi::NodalField{T}) where {ASS<:AbstractSysmatAssembler, T<:Number}

Compute the material stiffness matrix.

FEMMShellT3FFModule functions

FinEtools.FEMMBaseModule.associategeometry! — Method

associategeometry!(self::FEMMShellT3FF,  geom::NodalField{FFlt})

Associate geometry with the FEMM.

In this case it means evaluate the nodal normals.

FinEtools.FEMMBaseModule.inspectintegpoints — Method

inspectintegpoints(self::AbstractFEMMDeforLinear,
  geom::NodalField{FFlt},  u::NodalField{TI},
  dT::NodalField{FFlt},
  felist::FIntVec,
  inspector::F,  idat, quantity=:Cauchy;
  context...) where {TI<:Number, F<:Function}

Inspect integration point quantities.

geom - reference geometry field

u - displacement+rotation field

dT - temperature difference field

felist - indexes of the finite elements that are to be inspected: The fes to be included are: fes[felist].

context - structure: see the update!() method of the material.

inspector - function with the signature idat = inspector(idat, j, conn, x, out, loc); where idat - a structure or an array that the inspector may use to maintain some state, for instance minimum or maximum of stress, j is the element number, conn is the element connectivity, out is the output of the update!() method, loc is the location of the integration point in the reference configuration.

Return

The updated inspector data is returned.

FinEtoolsFlexStructures.FEMMShellT3FFModule.make — Method

make(integdomain, mcsys, material)

Make a T3FF FEMM from the integration domain, a coordinate system to define the orientation of the normals, and a material.

FinEtoolsFlexStructures.FEMMShellT3FFModule.make — Method

make(integdomain, mcsys, material)

Make a T3FF FEMM from the integration domain, and a material. Default isoparametric method for computing the normals is used.

FinEtoolsFlexStructures.FEMMShellT3FFModule.mass — Method

mass(self::FEMMShellT3FF,  assembler::A,  geom::NodalField{FFlt}, dchi::NodalField{TI}) where {A<:AbstractSysmatAssembler, TI<:Number}

Compute the diagonal (lumped) mass matrix

The mass matrix can be expected to be non-singular.

FinEtoolsFlexStructures.FEMMShellT3FFModule.num_normals — Method

num_normals(self::FEMMShellT3FF)

Compute the summary of the nodal normals.

FinEtoolsFlexStructures.FEMMShellT3FFModule.stiffness — Method

stiffness(self::FEMMShellT3FF, assembler::ASS, geom0::NodalField{FFlt}, u1::NodalField{TI}, Rfield1::NodalField{TI}, dchi::NodalField{TI}) where {ASS<:AbstractSysmatAssembler, TI<:Number}

Compute the material stiffness matrix.

FEMMPointMassModule functions

FinEtoolsFlexStructures.FEMMPointMassModule.mass — Method

mass(self::FEMMCorotBeam,  assembler::A,
  geom::NodalField{FFlt},
  u::NodalField{T}) where {A<:AbstractSysmatAssembler, T<:Number}

Compute the consistent mass matrix

This is a general routine for the abstract linear-deformation FEMM.

FEMMPointGroundedSpringModule functions

FinEtoolsFlexStructures.FEMMPointGroundedSpringModule.stiffness — Method

mass(self::FEMMCorotBeam,  assembler::A,
  geom::NodalField{FFlt},
  u::NodalField{T}) where {A<:AbstractSysmatAssembler, T<:Number}

Compute the stiffness matrix

RotUtilModule functions

FinEtoolsFlexStructures.RotUtilModule.initial_Rfield — Method

initial_Rfield(fens)

Set up initial rotation field (rotation matrices are all identities)

FinEtoolsFlexStructures.RotUtilModule.linear_update_rotation_field! — Method

linear_update_rotation_field!(Rfield, dchi)

Update rotation field by linear incremental rotation.

FinEtoolsFlexStructures.RotUtilModule.update_rotation_field! — Method

update_rotation_field!(Rfield, dchi)

Update rotation field by exponential incremental rotation.

VisUtilModule functions

FinEtoolsFlexStructures.VisUtilModule.default_layout_3d — Method

default_layout_3d(; kwargs...)

Set up the default layout for three-dimensional plots

FinEtoolsFlexStructures.VisUtilModule.plot_from_json — Method

plot_from_json(fn)

Retrieve plot from a JSON file and display it

FinEtoolsFlexStructures.VisUtilModule.plot_midline — Method

plot_midline(fens, fes; color = "rgb(155, 155, 255)", lwidth = 4)

Plot the midline

FinEtoolsFlexStructures.VisUtilModule.plot_midsurface — Method

plot_midline(fens, fes; color = "rgb(155, 155, 255)", lwidth = 4)

Plot the midsurface

FinEtoolsFlexStructures.VisUtilModule.plot_nodes — Method

plot_nodes(fens; color = "#1f77b4")

Plot nodes as markers

FinEtoolsFlexStructures.VisUtilModule.plot_points — Method

plot_points(xyz; color = "#1f77b4", msize = 4)

Plot points as markers

FinEtoolsFlexStructures.VisUtilModule.plot_solid — Method

plot_solid(fens, fes; kwargs...)

Plot beam structure as solid

FinEtoolsFlexStructures.VisUtilModule.plot_space_box — Method

plot_space_box(xyz; color = BACKGROUNDCOLOR)

Plot points to set up the global 3D box

FinEtoolsFlexStructures.VisUtilModule.plot_triads — Method

plot_triads(fens; kwargs...)

Plot the node triads.

FinEtoolsFlexStructures.VisUtilModule.render — Method

render(traces; kwargs...)

Render a plot

Key word arguments:

config = PlotConfig()

FinEtoolsFlexStructures.VisUtilModule.save_to_json — Method

save_to_json(pl, fn)

Save plot to a JSON file

FinEtoolsFlexStructures.VisUtilModule.space_aspectratio — Method

space_aspectratio(xyz; margin = 0.0)

Compute the aspect ratio of the space given by the array of locations

TransformerModule functions