Functions

Table of Contents

CrossSectionModule functions

CompositeLayupModule functions

FESetShellT3Module functions

MeshFrameMemberModule functions

FEMMCorotBeamModule functions

FEMMShellT3FFModule functions

FEMMShellT3FFCompModule functions

FEMMPointMassModule functions

FEMMPointGroundedSpringModule functions

RotUtilModule functions

VisUtilModule functions

TransformerModule functions

CrossSectionModule functions

CompositeLayupModule functions

FinEtoolsFlexStructures.CompositeLayupModule.cartesian_csys — Method

cartesian_csys(axes)

Create a material Cartesian coordinate system.

axes = tuple of signed labels of the axes. For instance, (1, 2, 3) creates a coordinate system identical to the global cartesian coordinate system. (2, 1, -3) creates a coordinate system so that the first material basis vector is along the second global basis vector, the second material basis vector is along the first global basis vector, and the third material basis vector is opposite to the third global basis vector.

FinEtoolsFlexStructures.CompositeLayupModule.lamina_material — Method

lamina_material(E1, E2, nu12, G12, G13, G23)

Create a transversely isotropic lamina material with default (zero) mass density.

FinEtoolsFlexStructures.CompositeLayupModule.lamina_material — Method

lamina_material(rho, E1, E2, nu12, G12, G13, G23)

Create a transversely isotropic lamina material.

FinEtoolsFlexStructures.CompositeLayupModule.lamina_material — Method

lamina_material(rho, E, nu)

Create an isotropic lamina material.

FinEtoolsFlexStructures.CompositeLayupModule.lamina_material — Method

lamina_material(E, nu)

Create an isotropic lamina material with default (zero) mass density.

FinEtoolsFlexStructures.CompositeLayupModule.laminate_stiffnesses! — Method

laminate_stiffnesses!(cl::CompositeLayup, A, B, C)

Compute the laminate stiffnesses, membrane, extension-bending coupling, and bending.

FinEtoolsFlexStructures.CompositeLayupModule.laminate_transverse_stiffness! — Method

laminate_transverse_stiffness!(cl::CompositeLayup, H)

Computed the laminate transverse stiffness.

FinEtoolsFlexStructures.CompositeLayupModule.plane_stress_T_matrix! — Method

plane_stress_T_matrix!( Tinvm::Array{T, 2},
 angle) where {T}

Compute the transformation matrix between strain tensor components on the layup coordinate system into the ply coordinate system.

angle = angle between the first basis vector of the layup coordinate system and the first basis vector of the ply coordinate system

FinEtoolsFlexStructures.CompositeLayupModule.plane_stress_T_matrix_eng! — Method

plane_stress_T_matrix_eng!( Tinvm::Array{T, 2},
 angle) where {T}

Compute the transformation matrix between strain engineering components on the layup coordinate system into the ply coordinate system.

angle = angle between the first basis vector of the layup coordinate system and the first basis vector of the ply coordinate system

The matrix of transformation is Tme = R * Tm / R, where Tm is the transformation matrix in tensor components, and R is the Reuter matrix, R = [1 0 0; 0 1 0; 0 0 2].

FinEtoolsFlexStructures.CompositeLayupModule.plane_stress_Tinv_matrix! — Method

plane_stress_Tinv_matrix!(Tinvm::Array{T, 2}, angle) where {T}

Compute the transformation matrix between strain tensor components on the ply coordinate system into the layout coordinate system.

angle = angle between the first basis vector of the layup coordinate system and the first basis vector of the ply coordinate system

FinEtoolsFlexStructures.CompositeLayupModule.plane_stress_Tinv_matrix_eng! — Method

plane_stress_Tinv_matrix!(Tinvm::Array{T, 2}, angle) where {T}

Compute the transformation matrix between strain engineering components on the ply coordinate system into the layout coordinate system.

angle = angle between the first basis vector of the layup coordinate system and the first basis vector of the ply coordinate system

The matrix of transformation is Tme = R * Tm / R, where Tm is the transformation matrix in tensor components, and R is the Reuter matrix, R = [1 0 0; 0 1 0; 0 0 2].

FinEtoolsFlexStructures.CompositeLayupModule.thickness — Method

thickness(cl::CompositeLayup)

Compute the thickness of the layup (some of the thicknesses of the plies).

FESetShellT3Module functions

MeshFrameMemberModule functions

FinEtoolsFlexStructures.MeshFrameMemberModule.frame_member — Method

frame_member(xyz, nL, crosssection; label = 0)

Mesh of a generally curved beam member given by the location of the vertices of the spline curve.

FinEtoolsFlexStructures.MeshFrameMemberModule.fuse_members — Method

fuse_members(members; tolerance = 0.001)

Fuse members by merging the meshes for the members.

Computes an array of meshes.

FinEtoolsFlexStructures.MeshFrameMemberModule.merge_members — Method

merge_members(members; tolerance = 0.001)

Merge together meshes of the members.

Uses fuse_members to merge the nodes, then concatenates the finite elements.

FEMMCorotBeamModule functions

FinEtoolsFlexStructures.FEMMCorotBeamModule.distribloads_global — Method

distribloads_global(self::FEMMCorotBeam, geom0::NodalField{FFlt}, u1::NodalField{T}, Rfield1::NodalField{T}, dchi::NodalField{T}, fi) where {T<:Number}

Compute the load vector due to distributed loads.

Compute the global load vector corresponding to applied distributed load. Here it means force per unit length of the beam, in the configuration u1,Rfield1.

Note: the force intensity must be uniform across the entire element. Note: the force intensity is given in the global coordinates.

FinEtoolsFlexStructures.FEMMCorotBeamModule.geostiffness — Method

geostiffness(self::FEMMCorotBeam, assembler::ASS, geom0::NodalField{FFlt}, u1::NodalField{T}, Rfield1::NodalField{T}, dchi::NodalField{T}) where {ASS<:AbstractSysmatAssembler, T<:Number}

Compute the geometric stiffness matrix.

FinEtoolsFlexStructures.FEMMCorotBeamModule.gyroscopic — Method

gyroscopic(self::FEMMCorotBeam,  assembler::A,
  geom::NodalField{FFlt},
  u::NodalField{T}) where {A<:AbstractSysmatAssembler, T<:Number}

Compute the quadratic-inertial-term (gyroscopic) mass matrix

This is a general routine for the abstract linear-deformation FEMM.

FinEtoolsFlexStructures.FEMMCorotBeamModule.mass — Method

mass(self::FEMMCorotBeam,  assembler::A,
  geom::NodalField{FFlt},
  u::NodalField{T}) where {A<:AbstractSysmatAssembler, T<:Number}

Compute the consistent mass matrix

This is a general routine for the abstract linear-deformation FEMM.

FinEtoolsFlexStructures.FEMMCorotBeamModule.restoringforce — Method

restoringforce(self::FEMMCorotBeam, assembler::ASS, geom0::NodalField{FFlt}, u1::NodalField{T}, Rfield1::NodalField{T}, dchi::NodalField{T}) where {ASS<:AbstractSysmatAssembler, T<:Number}

Compute the vector of the restoring elastic forces

FinEtoolsFlexStructures.FEMMCorotBeamModule.stiffness — Method

stiffness(self::FEMMCorotBeam, assembler::ASS, geom0::NodalField{FFlt}, u1::NodalField{T}, Rfield1::NodalField{T}, dchi::NodalField{T}) where {ASS<:AbstractSysmatAssembler, T<:Number}

Compute the material stiffness matrix.

FEMMShellT3FFModule functions

FinEtools.FEMMBaseModule.associategeometry! — Method

associategeometry!(self::FEMMShellT3FF,  geom::NodalField{FFlt})

Associate geometry with the FEMM.

In this case it means evaluate the nodal normals.

FinEtools.FEMMBaseModule.inspectintegpoints — Method

inspectintegpoints(self::AbstractFEMMDeforLinear,
  geom::NodalField{FFlt},  u::NodalField{TI},
  dT::NodalField{FFlt},
  felist::FIntVec,
  inspector::F,  idat, quantity=:Cauchy;
  context...) where {TI<:Number, F<:Function}

Inspect integration point quantities.

geom - reference geometry field

u - displacement+rotation field

dT - temperature difference field

felist - indexes of the finite elements that are to be inspected: The fes to be included are: fes[felist].

context - structure: see the update!() method of the material.

inspector - function with the signature idat = inspector(idat, j, conn, x, out, loc); where idat - a structure or an array that the inspector may use to maintain some state, for instance minimum or maximum of stress, j is the element number, conn is the element connectivity, out is the output of the update!() method, loc is the location of the integration point in the reference configuration.

Return

The updated inspector data is returned.

FinEtoolsFlexStructures.FEMMShellT3FFModule.make — Method

make(integdomain, mcsys, material)

Make a T3FF FEMM from the integration domain, a coordinate system to define the orientation of the normals, and a material.

FinEtoolsFlexStructures.FEMMShellT3FFModule.make — Method

make(integdomain, mcsys, material)

Make a T3FF FEMM from the integration domain, and a material. Default isoparametric method for computing the normals is used.

FinEtoolsFlexStructures.FEMMShellT3FFModule.mass — Method

mass(self::FEMMShellT3FF,  assembler::A,  geom::NodalField{FFlt}, dchi::NodalField{TI}) where {A<:AbstractSysmatAssembler, TI<:Number}

Compute the diagonal (lumped) mass matrix

The mass matrix can be expected to be non-singular.

FinEtoolsFlexStructures.FEMMShellT3FFModule.num_normals — Method

num_normals(self::FEMMShellT3FF)

Compute the summary of the nodal normals.

FinEtoolsFlexStructures.FEMMShellT3FFModule.stiffness — Method

stiffness(self::FEMMShellT3FF, assembler::ASS, geom0::NodalField{FFlt}, u1::NodalField{TI}, Rfield1::NodalField{TI}, dchi::NodalField{TI}) where {ASS<:AbstractSysmatAssembler, TI<:Number}

Compute the material stiffness matrix.

FEMMShellT3FFCompModule functions

FinEtools.FEMMBaseModule.associategeometry! — Method

associategeometry!(self::FEMMShellT3FFComp,  geom::NodalField{FFlt})

Associate geometry with the FEMM.

In this case it means evaluate the nodal normals.

FinEtools.FEMMBaseModule.inspectintegpoints — Method

inspectintegpoints(self::AbstractFEMMDeforLinear,
  geom::NodalField{FFlt},  u::NodalField{TI},
  dT::NodalField{FFlt},
  felist::FIntVec,
  inspector::F,  idat, quantity=:Cauchy;
  context...) where {TI<:Number, F<:Function}

Inspect integration point quantities.

geom - reference geometry field

u - displacement+rotation field

dT - temperature difference field

felist - indexes of the finite elements that are to be inspected: The fes to be included are: fes[felist].

context - structure: see the update!() method of the material.

inspector - function with the signature idat = inspector(idat, j, conn, x, out, loc); where idat - a structure or an array that the inspector may use to maintain some state, for instance minimum or maximum of stress, j is the element number, conn is the element connectivity, out is the output of the update!() method, loc is the location of the integration point in the reference configuration.

Return

The updated inspector data is returned.

FinEtoolsFlexStructures.FEMMShellT3FFCompModule.make — Method

make(integdomain, layup)

Make a T3FFComp FEMM from the integration domain and a composite layup.

FinEtoolsFlexStructures.FEMMShellT3FFCompModule.mass — Method

mass(self::FEMMShellT3FFComp,  assembler::A,  geom::NodalField{FFlt}, dchi::NodalField{TI}) where {A<:AbstractSysmatAssembler, TI<:Number}

Compute the diagonal (lumped) mass matrix

The mass matrix can be expected to be non-singular.

FinEtoolsFlexStructures.FEMMShellT3FFCompModule.num_normals — Method

num_normals(self::FEMMShellT3FFComp)

Compute the summary of the nodal normals.

FinEtoolsFlexStructures.FEMMShellT3FFCompModule.stiffness — Method

stiffness(self::FEMMShellT3FFComp, assembler::ASS, geom0::NodalField{FFlt}, u1::NodalField{TI}, Rfield1::NodalField{TI}, dchi::NodalField{TI}) where {ASS<:AbstractSysmatAssembler, TI<:Number}

Compute the material stiffness matrix.

FEMMPointMassModule functions

FinEtoolsFlexStructures.FEMMPointMassModule.mass — Method

mass(self::FEMMCorotBeam,  assembler::A,
  geom::NodalField{FFlt},
  u::NodalField{T}) where {A<:AbstractSysmatAssembler, T<:Number}

Compute the consistent mass matrix

This is a general routine for the abstract linear-deformation FEMM.

FEMMPointGroundedSpringModule functions

FinEtoolsFlexStructures.FEMMPointGroundedSpringModule.stiffness — Method

mass(self::FEMMCorotBeam,  assembler::A,
  geom::NodalField{FFlt},
  u::NodalField{T}) where {A<:AbstractSysmatAssembler, T<:Number}

Compute the stiffness matrix

RotUtilModule functions

FinEtoolsFlexStructures.RotUtilModule.initial_Rfield — Method

initial_Rfield(fens)

Set up initial rotation field (rotation matrices are all identities)

FinEtoolsFlexStructures.RotUtilModule.update_rotation_field! — Method

update_rotation_field!(Rfield, dchi)

Update rotation field by exponential incremental rotation.

VisUtilModule functions

Modules = [Examples.VisUtilModule]
Private = true
Order = [:function]

TransformerModule functions