Types

Table of Contents

CompositeLayupModule types

CrossSectionModule types

MeshFrameMemberModule types

RotUtilModule types

TransformerModule types

FESetL2BeamModule types

FESetShellT3Module types

FEMMCorotBeamModule types

FEMMLinBeamModule types

FEMMRITBeamModule types

FEMMShellT3FFModule types

FEMMShellT3FFCompModule types

FEMMPointMassModule types

FEMMPointGroundedSpringModule types

CompositeLayupModule types

FinEtoolsFlexStructures.CompositeLayupModule.CompositeLayup — Type

CompositeLayup

Type for composite layup.

FinEtoolsFlexStructures.CompositeLayupModule.CompositeLayup — Method

CompositeLayup(name, plies, mcsys)

Create a composite layup.

Provide the name, the array of plies, and the coordinate system that defines the orientation of the composite layup. The first base spector of this coordinate system is the reference direction for the layup.

FinEtoolsFlexStructures.CompositeLayupModule.Ply — Type

Ply{M} <: AbstractPly

Type of a ply.

FinEtoolsFlexStructures.CompositeLayupModule.Ply — Method

Ply{M} <: AbstractPly

Create a ply.

Provide name, material of the ply, thickness of the ply, an angle between the first bases vector of the layup coordinate system and the first direction of the ply material coordinate system.

CrossSectionModule types

FinEtoolsFlexStructures.CrossSectionModule.CrossSectionCircle — Method

CrossSectionCircle(radius, x1x2_vector, shear_resisting_area_factor = 6/7; label = 0)

Create a circular section for use with a Timoshenko model.

Commonly accepted value is shear_resisting_area_factor = 6/7.

FinEtoolsFlexStructures.CrossSectionModule.CrossSectionCircle — Method

CrossSectionCircle(radius, x1x2_vector; label = 0)

Create a circular section for use with a Bernoulli model.

FinEtoolsFlexStructures.CrossSectionModule.CrossSectionRectangle — Method

CrossSectionRectangle(d2f, d3f, x1x2_vector; label = 0)

Create a rectangular section for use with a Timoshenko model.

Commonly accepted value is shear_resisting_area_factor = 5/6.

FinEtoolsFlexStructures.CrossSectionModule.CrossSectionRectangle — Method

CrossSectionRectangle(d2f, d3f, x1x2_vector; label = 0)

Create a rectangular section for use with a Bernoulli model.

MeshFrameMemberModule types

RotUtilModule types

TransformerModule types

FinEtoolsFlexStructures.TransformerModule.Layup2ElementAngle — Type

Layup2ElementAngle{T}

A callable object: computes the cosine and sine of the angle between the element frame and the layup reference direction.

FinEtoolsFlexStructures.TransformerModule.TransformerQtEQ — Type

TransformerQtEQ

TransformerQtEQ of element matrices.

A callable object: computes Q^T * E * Q, where E the element stiffness matrix, and Q is the transformation matrix. Both are assumed to be square.

Buffers the intermediate result. Hence no allocation is incurred.

FinEtoolsFlexStructures.TransformerModule.TransformerQtEQ — Method

(o::TransformerQtEQ)(E, Q)

Perform the transformation on the matrix E with the transformation matrix Q: Ebar = Q^T * E * Q.

FESetL2BeamModule types

FinEtoolsFlexStructures.FESetL2BeamModule.FESetL2Beam — Type

mutable struct FESetL2Beam{CT} <: AbstractFESet1Manifold{2}

Type of a two-node beam finite element set.

The elements in the set have the same cross section type, but since the cross section dimensions may depend on the coordinate, each individual element may have a different area, second moments of area, etc.

FinEtoolsFlexStructures.FESetL2BeamModule.FESetL2Beam — Method

FESetL2Beam(N::IT, crosssection::CT) where {IT<:Integer, CT}

Constructor.

Supply the total number of elements in the set, and the cross-section definition. The parameters of the section are assumed to be uniform along the set of the beam elements.

FESetShellT3Module types

FinEtoolsFlexStructures.FESetShellT3Module.FESetShellT3 — Type

struct FESetShellT3 <: AbstractFESet2Manifold{3}

Type of a finite element set for the three-node shell.

FEMMCorotBeamModule types

FinEtoolsFlexStructures.FEMMCorotBeamModule.FEMMCorotBeam — Type

mutable struct FEMMCorotBeam{ID<:IntegDomain{S} where {S<:FESetL2}, M<:MatDeforElastIso} <: AbstractFEMM

Type for co-rotational beam finite element modeling machine (FEMM).

FinEtoolsFlexStructures.FEMMCorotBeamModule.FEMMCorotBeam — Method

FEMMCorotBeam(integdomain::IntegDomain{S, F}, material::MatDeforElastIso) where {S<:FESetL2, F<:Function}

Constructor.

Supply the integration domain and the material.

FEMMLinBeamModule types

FinEtoolsFlexStructures.FEMMLinBeamModule.FEMMLinBeam — Type

mutable struct FEMMLinBeam{ID<:IntegDomain{S} where {S<:FESetL2}, M<:MatDeforElastIso} <: AbstractFEMM

Type for linear beam finite element modeling machine.

The beam can be connected to the nodes with given eccentricities (the transverse eccentricity is uniform along its length). Only linear kinematics is implemented at the moment. The beam stiffness can be either shear-flexible (Timoshenko), or shear-rigid (Bernoulli-Euler). The local beam stiffness is expressed analytically: no numerical integration is involved.

FinEtoolsFlexStructures.FEMMLinBeamModule.FEMMLinBeam — Method

FEMMLinBeam(
    integdomain::ID,
    material::M,
    uniform_eccentricity,
) where {ID<:IntegDomain{S} where {S<:FESetL2}, M<:MatDeforElastIso}

Constructor.

Supply the integration domain, material, and the eccentricity parameters (given in the order: first node, f1 direction, second node, f1 direction, f2 direction, f3 direction).

FinEtoolsFlexStructures.FEMMLinBeamModule.FEMMLinBeam — Method

FEMMLinBeam(
    integdomain::ID,
    material::M,
) where {ID<:IntegDomain{S} where {S<:FESetL2}, M<:MatDeforElastIso}

Constructor.

Supply the integration domain and the material. The eccentricities are assumed to be zero.

FEMMRITBeamModule types

FinEtoolsFlexStructures.FEMMRITBeamModule.FEMMRITBeam — Type

mutable struct FEMMRITBeam{ID<:IntegDomain{S} where {S<:FESetL2}, M<:MatDeforElastIso} <: AbstractFEMM

Type for linear reduced-integration beam finite element modeling machine.

Only linear kinematics is implemented at the moment, and only linear basis functions are available (i.e. it is a two-node element). The beam stiffness is shear-flexible(Timoshenko). The local beam stiffness is expressed analytically: the one-point numerical integration is hardwired.

FinEtoolsFlexStructures.FEMMRITBeamModule.FEMMRITBeam — Method

FEMMRITBeam(
    integdomain::ID,
    material::M,
) where {ID<:IntegDomain{S} where {S<:FESetL2}, M<:MatDeforElastIso}

Constructor.

Supply integration domain (mesh) and the material.

FEMMShellT3FFModule types

FinEtoolsFlexStructures.FEMMShellT3FFModule.FEMMShellT3FF — Type

mutable struct FEMMShellT3FF{ID<:IntegDomain{S} where {S<:FESetT3}, CS<:CSys, M} <: AbstractFEMM

Type for the finite element modeling machine of the T3 triangular Flat-Facet shell with the Discrete Shear Gap technology and a consistent handling of the normals.

With averaging of the transverse strain-displacement matrix or averaging of the transverse shear stiffness matrix to provide isotropic transverse shear response.

Also, the formulation is developed to correctly handle the coupling of twisting moments and transverse shear (such as in the twisted beam or the Raasch hook problems) by incorporating "nodal" normals.

Some of the programming developed consistently with the paper

[1] Cui et al, Analysis of plates and shells using an edge-based smoothed finite element method, Comput Mech (2010) 45:141–156 DOI 10.1007/s00466-009-0429-9

The stabilization factor of the shear term of

[2] Mikko Lyly, Rolf Stenberg and Teemu Vihinen, A stable bilinear element for the Reissner-Mindlin plate model Computer Methods in Applied Mechanics and Engineering 110 (1993) 343-357

is incorporated. Refer to expressions (3.12) and (3.13).

The treatment of the transformation between the element and nodal coordinates is carried out using a clean alternative to the publication

[3] Finite Elements in Analysis and Design 30 (1998) 235—242 The treatment of shell normals in ﬁnite element analysis Richard H. MacNeal, Charles T. Wilson, Robert L. Harder, Claus C. Hoﬀ The MacNeal-Schwendler Corporation, 815 Colorado Blvd., Los Angeles, CA 90041, USA

The formula for the element to nodal basis transformation is derived from the expression

[theta]_n = [A]_E^T [theta]_e + [A_3,1:3]_E^T [alpha_3]_e

by disconnecting the drilling degree of freedom from the bending degrees of freedom in the nodal basis.

The following features are incorporated to deal with nodal normals:

Nodal normals are averages of the normals of elements that meet at a node.

A crease in the surface is taken into account. In that case the normal are not averaged across the crease. At the nodes along the crease every element uses the normal to its surface instead of the nodal normal.

Configuration:

These attributes of the FEMM can be set after it's been created.

transv_shear_formulation: which formulation for the transverse shear stiffness?

FEMMShellT3FFModule.__TRANSV_SHEAR_FORMULATION_AVERAGE_B - averaged strains (default)

FEMMShellT3FFModule.__TRANSV_SHEAR_FORMULATION_AVERAGE_K - averaged stiffness

drilling_stiffness_scale: multiplier of the generalized stiffness coefficient

threshold_angle: angle in degrees. If a nodal normal subtends angle bigger then this threshold, the nodal normal at that note is marked as invalid.

mult_el_size: multiplier of the square of the element size, used to control transverse shear stiffness.

FinEtoolsFlexStructures.FEMMShellT3FFModule.FEMMShellT3FF — Method

FEMMShellT3FF(
    integdomain::ID,
    mcsys::CSys,
    material::M,
) where {IntegDomain{S} where {S<:FESetT3}, CS<:CSys, M}

Constructor of the T3FF shell FEMM.

FinEtoolsFlexStructures.FEMMShellT3FFModule.FEMMShellT3FF — Method

FEMMShellT3FF(
    integdomain::ID,
    material::M,
) where {ID<:IntegDomain{S} where {S<:FESetT3}, M}

Constructor of the T3FF shell FEMM.

FEMMShellT3FFCompModule types

FinEtoolsFlexStructures.FEMMShellT3FFCompModule.FEMMShellT3FFComp — Type

mutable struct FEMMShellT3FFComp{ID<:IntegDomain{S} where {S<:FESetT3}} <: AbstractFEMM

Type for the finite element modeling machine of the T3 triangular Flat-Facet shell with the Discrete Shear Gap technology and a consistent handling of the normals. This formulation is suitable for modelling of COMPOSITE (layered) materials.

For details for the homogeneous-shell refer to FEMMShellT3FF.

FinEtoolsFlexStructures.FEMMShellT3FFCompModule.FEMMShellT3FFComp — Method

FEMMShellT3FFComp(
    integdomain::ID,
    layup::CompositeLayup,
)  where {ID<:IntegDomain{S} where {S<:FESetT3}}

Constructor of the T3FFComp shell FEMM. All elements use a single layup.

FEMMPointMassModule types

FinEtoolsFlexStructures.FEMMPointMassModule.FEMMPointMass — Type

mutable struct FEMMPointMass{ID<:IntegDomain} <: AbstractFEMM

Type for linear deformation finite element modeling machine.

FEMMPointGroundedSpringModule types

FinEtoolsFlexStructures.FEMMPointGroundedSpringModule.FEMMPointGroundedSpring — Type

mutable struct FEMMPointGroundedSpring{ID<:IntegDomain} <: AbstractFEMM

Type for finite element modeling machine for grounded springs.